start refactor reportView update SetNewPassword write  reacteToEmaillView
