+++ lib/provider/user_provider/user_administration/user_administration._provider.dart
@@ -1,3 +1,4 @@
+import 'dart:convert';
 import 'dart:developer';
 
 import 'package:dio/dio.dart';
@@ -6,6 +7,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../../../constants/api/api.dart';
 import '../../../models/users_models/user_data_short/user_short.dart';
 import '../../../models/users_models/user_role/user_role.dart';
+import '../user_provider.dart';
 
 final userAdministrationProvider = NotifierProvider<UserAdministrationNotifer, List<UserDataShort>>(
     () => UserAdministrationNotifer());
@@ -29,6 +31,8 @@ class UserAdministrationNotifer extends Notifier<List<UserDataShort>> {
       'roles': [role.name],
     };
     try {
+      log(ref.watch(userProvider).userToken);
+      log(jsonEncode(newUser));
       final response = await _api.postCreateNewUser(newUser);
       if (response.statusCode != 200) {
         throw Exception('${response.statusCode} Invalid Api call ${response.data}');
