+++ lib/provider/user_provider/user_provider.dart
@@ -1,5 +1,6 @@
 import 'dart:developer';
 
+import 'package:dio/dio.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../../constants/api/api.dart';
 import '../../models/users_models/user_vm/user_vm.dart';
@@ -26,21 +27,19 @@ class UserNotifier extends Notifier<UserVM> {
     _api.deleteToken();
   }
 
-  Future<String?> getUserToken() async => _api.getToken;
-
   Future<bool> loginUser({
-    required String passwort,
+    required String password,
     required String userName,
     String? mandatID,
   }) async {
     final Map<String, dynamic> json = {
       'username': userName,
-      'password': passwort,
+      'password': password,
       'mandant': mandatID ?? '1',
     };
     try {
       final response = await _api.postloginUser(json);
-      if (response.statusCode == 401) {
+      if (response.statusCode != 200) {
         log('user not authorized');
         return false;
       }
@@ -48,7 +47,6 @@ class UserNotifier extends Notifier<UserVM> {
         log(response.data.toString());
         final data = (response.data as Map);
         final userToken = data.values.first as String;
-        // TODO: when token Exist load user with Token
         final newUser = state.copyWith(userToken: userToken);
         _api.storeToken(userToken);
 
@@ -68,19 +66,43 @@ class UserNotifier extends Notifier<UserVM> {
 
   /// TODO: This Api calls not Users!!! and is deprecated.
   void loadUsers() async {
-    final response = await _api.getProjectsDM;
-    if (response.statusCode == 200) {
-      final data = response.data;
-      final x = data.map((e) => UserVM.fromJson(data));
-      state = x;
-      return;
-    }
-    if (response.statusCode != 200) {
-      log('something went wrong by loadingUser ${response.data}');
-      return;
-    }
-    try {} catch (e) {
+    try {
+      final response = await _api.getProjectsDM;
+      if (response.statusCode == 200) {
+        final data = response.data;
+        final x = data.map((e) => UserVM.fromJson(data));
+        state = x;
+        return;
+      }
+      if (response.statusCode != 200) {
+        log('something went wrong by loadingUser ${response.data}');
+        return;
+      }
+    } catch (e) {
       throw Exception(e);
+    }
+  }
+
+  Future<bool> setNewPassword(String username, String oTP, String newPassword) async {
+    final json = {
+      'userName': username,
+      'oldPassword': oTP,
+      'newPassword': newPassword,
+    };
+    try {
+      final response = await _api.putSetNewPassword(json);
+      if (response.statusCode != 200) {
+        throw Exception(
+            'something went wrong by setNewPassword: status-> ${response.statusCode}\n${response.data}');
+      }
+      log(response.data);
+      return true;
+    } on DioException catch (e) {
+      log('${e.message}');
+      return false;
+    } catch (e) {
+      log(e.toString());
+      return false;
     }
   }
 }
