+++ lib/routes/app_routes.dart
@@ -3,7 +3,8 @@ import 'package:flutter/widgets.dart';
 
 import '../view/login_view/forgot_password_view.dart';
 import '../view/login_view/login_view.dart';
-import '../view/login_view/reset_password.dart';
+import '../view/login_view/react_to_email_view.dart';
+import '../view/login_view/set_new_password_view.dart';
 import '../view/main_navigation_view/main_nav_view.dart';
 import '../view/start_view/start_view.dart';
 
@@ -13,12 +14,14 @@ class AppRoutes {
   static const setPasswordScreen = '/set_password_view';
   static const viewScreen = '/view_view';
   static const forgotPassword = '/forgot_password';
+  static const reactOnPwResetView = 'request_pw_reset';
 
   static Map<String, WidgetBuilder> routes = {
     anmeldeScreen: (context) => const LoginView(),
     forgotPassword: (context) => const ForgetScreen(),
     initialRoute: (context) => const StartView(),
-    setPasswordScreen: (context) => const PasswordView(),
+    setPasswordScreen: (context) => const SetNewPasswordView(),
     viewScreen: (context) => const MainViewNavigator(),
+    reactOnPwResetView: (context) => const ReactToEmailView(),
   };
 }
